refactor: icons 파일 분리
